[BE] Refactor: 보관함 트랙 서비스 분리 및 적용
